Changed the -O3 option to -O2 for avoiding functions inlining.
The -O3 option results the following assembly around switch_context calling
at the resume function in src/coroutine.c. This is because that the -O3 option
includes -finline-functions.

      .loc 1 163 0
      addl $24, %eax
      addl $36, %edi
      movl %eax, %esi
  .LVL51:
  #APP
      movl $1f, (%edi)
      movl %esp, 4(%edi)
      movl %ebp, 8(%edi)
      movl 8(%esi), %ebp
      movl 4(%esi), %esp
      jmp *(%esi)
  1:

  .LVL52:
  #NO_APP
  .LBE60:
  .LBE59:
       .loc 1 421 0
       movl -208(%ebp), %edi
       movl 8(%ebp), %eax
       movl 64(%edi), %edx
       call return_args

"jmp *(%esi)" resumes a coroutine, so that this is a kind of subroutine calling.
ebp register will be changed in coroutines, but this code doesn't store/restore
a value of ebp register. So that, "movl -208(%ebp), %edi" will refer invalid
ebp register.