Fixed run_command builtins function to read stdout and stderr of child processes after Process#poll returned not nil.
In a past version, run_command used Process#wait and attempted to read all
output after calling Process#wait. But in this implementation, a child process
might be locked in writing stdout/stderr. So the parent process (Yog) must read
stdout/stderr before the child process's terminating. The following change is
due to this reason.

  $ git diff -r db59bbba7216b459e8567f9811183d2c5e24887c -r 4f82a8ed1a9748cd054606470a93661e2fcd1511 src/builtins.yg
  diff --git a/src/builtins.yg b/src/builtins.yg
  index d6b57be..671a4f2 100644
  --- a/src/builtins.yg
  +++ b/src/builtins.yg
  @@ -6,9 +6,12 @@ def run_command(*cmd, &block=nil)
       block(proc.stdin)
     end
     proc.stdin.close()
  -  status = proc.wait()
  -  stdout = proc.stdout.read()
  -  stderr = proc.stderr.read()
  +  stdout = ""
  +  stderr = ""
  +  while (status = proc.poll()) == nil
  +    stdout << proc.stdout.read()
  +    stderr << proc.stderr.read()
  +  end
     return status, stdout, stderr
   end

In this case, I understood that a child process cannot terminate (return non-nil
value at Process#poll) before the parent process read all of stdout/stderr. But
it seems false. Even if Process#poll returns non-nil, a child process holds
stdout/stderr to be read. This commit is for this phenomenon.