Fixed testcase.TestLib class to be enabled only for yog-generational.
This setting was the following at first;

  from os.path import basename, exists, join, splitext

  ...(snip)...

  skipif = pytest.mark.skipif

  @skipif("splitext(basename(get_command()))[0] != \"yog-generational\"")
  class TestLib(TestCase):
      pass

But the expression in the skipif decorator seems to be evaluated in the scope
where the child classes are in. testcase.TestLib is delivered at test_socket.py
and test_zlib.py. These test_{socket,zlib}.py doesn't include the definition of
split and so on. So if we run py.test, these scripts became ERROR, and reported
the following;

  Error evaluating 'skipif' expression
      splitext(basename(get_command()))[0] != "yog-generational"
      NameError: name 'splitext' is not defined