Fixed Makefiles in ext to work "make install".
When I built Yog with Yog-0.0.10.tar.xz in FreeBSD 8.1, mkdir_p in
ext/**/Makefile.in is defined as "$(SHELL) $(mkinstalldirs)". Makefiles did not
have definitions of SHELL and mkinstalldirs, so "make install" failed.

This depends on configure. configure in Yog-0.0.10.tar.xz was generated by
Autoconf 2.63, but I used Autoconf 2.67 in FreeBSD 8.1 (I don't know that this
difference is a true reason).

I guess by my shell history that I did not "make pkgcheck" in FreeBSD 8.1 for
0.0.10 release.